Document

+++ Controller/PositionsController.php
@@ -4,94 +4,100 @@ class PositionsController extends AppController
     var $name = 'Positions';
     var $layout = 'default';
 
-    function beforeFilter()
-    {
+    //Check User login
+    function beforeFilter(){
         parent::beforeFilter();
     }
 
-    function index_intra()
-    {
-        if($this->Session->read('Auth.User.role') != "intra")
-        {
-            die('You are not authorized ');
+    //INTRA view redirect
+    function index_intra(){
+        if($this->Session->read('Auth.User.role') != "intra") {
+            throw new NotFoundException();
         }
-        else
-        {
+        else {
             $this->set('positions', $this->Position->find('all'));
             $this->set('title_for_layout', 'Positions');
         }
     }
 
-    function index_student()
-    {
-        if($this->Session->read('Auth.User.role') != "student")
-        {
-            die('You are not authorized ');
+    //Student view redirect
+    function index_student(){
+        if($this->Session->read('Auth.User.role') != "student") {
+            throw new NotFoundException();
         }
-        else
-        {
+        else {
             $this->set('positions', $this->Position->find('all', array('conditions'=>array('Position.Active'=>'1'))));
-            $this->set('title_for_layout', 'Positions');        }
+            $this->set('title_for_layout', 'Positions');
+        }
     }
 
-    function index_employer()
-    {
-        if($this->Session->read('Auth.User.role') != "employer")
-        {
-            die('You are not authorized ');
+    //Employer view redirect
+    function index_employer(){
+        if($this->Session->read('Auth.User.role') != "employer") {
+            throw new NotFoundException();
         }
-        else
-        {
-            $this->set('positions', $this->Position->find('all', array('conditions'=>array('Position.Parent'=>$this->Session->read('Auth.User.id')))));
-            $this->set('title_for_layout', 'Positions');        }
+        else {
+            $this->set('positions', $this->Position->find('all',
+                array('conditions'=>array('Position.Parent'=>$this->Session->read('Auth.User.id')))));
+            $this->set('title_for_layout', 'Positions');
+        }
     }
 
-    function index_support()
-    {
-        if($this->Session->read('Auth.User.role') != "support")
-        {
-            die('You are not authorized ');
+    //Support view redirect
+    function index_support(){
+        if($this->Session->read('Auth.User.role') != "support") {
+            throw new NotFoundException();
         }
-        else
-        {
+        else {
             $this->set('positions', $this->Position->find('all'));
-            $this->set('title_for_layout', 'Positions');        }
+            $this->set('title_for_layout', 'Positions');
+        }
     }
 
+    //View function
     function view($ID = NULL)
     {
         $user = $this->Auth->user();
-        //if(!$this->Acl->check($user['role'], 'Position', 'view'))
-            //if(!$this->Access->check('User', 'view'))
-        //{
-        //    die('You are not authorized ');
-       // }
-       // else {
+        if(!$this->Acl->check($user['role'], 'Position', 'view')) {
+            throw new NotFoundException();
+        }
+        else {
             $this->set('position', $this->Position->read(NULL, $ID));
-
-        // }
+        }
     }
 
-    function add()
+    //View student function , used for PDF
+    function view_student($ID = NULL)
     {
         $user = $this->Auth->user();
-        if(!$this->Acl->check($user['role'], 'Position', 'create'))
-            //if(!$this->Access->check('User', 'view'))
-        {
-            die('You are not authorized ');
+        if(!$this->Acl->check($user['role'], 'Position', 'view')) {
+            throw new NotFoundException();
+        }
+        else {
+        $this->set('position', $this->Position->read(NULL, $ID));
+        }
+    }
+
+    //Add function
+    function add() {
+        $user = $this->Auth->user();
+        if(!$this->Acl->check($user['role'], 'Position', 'create')) {
+            throw new NotFoundException();
         }
         else {
             if (!empty($this->data)) {
                 if ($this->Position->save($this->data)) {
                     $this->Session->setFlash('Position sucessfuly added');
                     $this->redirect(array('position' => 'index'));
-                } else {
+                }
+                else {
                     $this->Session->setFlash('Position unsucessfuly added');
                 }
             }
         }
     }
 
+    //Edit function
     function edit($ID = NULL)
     {
         if (!$ID) {
@@ -120,21 +126,22 @@ class PositionsController extends AppController
         }
     }
 
+    //Delete function
     function delete($ID = NULL)
     {
         $user = $this->Auth->user();
-        //if(!$this->Acl->check($user['role'], 'Position', 'delete'))
-            //if(!$this->Access->check('User', 'view'))
-        //{
-            //die('You are not authorized ');
-        //}
-        //else {
+        if(!$this->Acl->check($user['role'], 'Position', 'delete'))
+        {
+            throw new NotFoundException();
+        }
+        else {
             $this->Position->delete($ID);
             $this->Session->setFlash('Position deleted');
             $this->redirect(array('action' => 'index'));
-        //}
+        }
     }
 
+    //Latest positions added
     public function latest() {
         if (empty($this->request->params['requested'])) {
             throw new ForbiddenException();
